Refactor model imports to use explicit submodule paths
Import model classes directly from their respective modules (cnn, mlp)
rather than through the models package __init__.py re-exports.
