+++ experiments/centralised/hidden_vulnerability_icml_2018/experiment_1.py
@@ -9,7 +9,8 @@ from krum.primitives.aggregators.brute import Brute
 from krum.primitives.aggregators.geomed import GeoMed
 from krum.primitives.aggregators.krum import Krum
 from krum.primitives.attacks.small_perturbation import SmallPerturbationAttack
-from krum.primitives.models import Krum2017CNN, Krum2017MLPMnist
+from krum.primitives.models.cnn import Krum2017CNN
+from krum.primitives.models.mlp import Krum2017MLPMnist
 
 from .plot import _LABEL_KEYS_EXP1, _LEGEND_STYLE_EXP1, plot_accuracy_curves
 from .run import hidden_vulnerability_experiment
