+++ experiments/centralised/hidden_vulnerability_icml_2018/experiment_2.py
@@ -15,7 +15,8 @@ from krum.primitives.aggregators.bulyan import Bulyan
 from krum.primitives.aggregators.geomed import GeoMed
 from krum.primitives.aggregators.krum import Krum
 from krum.primitives.attacks.small_perturbation import SmallPerturbationAttack
-from krum.primitives.models import Krum2017CNN, Krum2017MLPMnist
+from krum.primitives.models.cnn import Krum2017CNN
+from krum.primitives.models.mlp import Krum2017MLPMnist
 
 from .plot import _LABEL_KEYS_EXP2, _LEGEND_STYLE_EXP2, plot_accuracy_curves
 from .run import hidden_vulnerability_experiment
