+++ experiments/centralised/hidden_vulnerability_icml_2018/experiment_3.py
@@ -6,7 +6,8 @@ Bulyan cost vs mini-batch size. No actual attack (f=0).
 from krum.orchestration import Orchestrator
 from krum.primitives.aggregators.average import Average
 from krum.primitives.aggregators.bulyan import Bulyan
-from krum.primitives.models import Krum2017CNN, Krum2017MLPMnist
+from krum.primitives.models.cnn import Krum2017CNN
+from krum.primitives.models.mlp import Krum2017MLPMnist
 
 from .plot import _LABEL_KEYS_EXP3, _LEGEND_STYLE_EXP3, plot_accuracy_vs_batch
 from .run import hidden_vulnerability_experiment
