+++ experiments/centralised/hidden_vulnerability_icml_2018/short_experiment.py
@@ -11,7 +11,7 @@ from krum.primitives.attacks.alie import (
     ALIEAttack,
     Direction,
 )
-from krum.primitives.models import Krum2017MLPMnist
+from krum.primitives.models.mlp import Krum2017MLPMnist
 
 from .run import hidden_vulnerability_experiment
